Organize CLI files and tests by responsibility
Group argument parsing, invocation dispatch, command handlers and output handling in named files. Fold the analysis selector into check execution and place compatibility, config, help, output and completion tests with their subjects.
