+++ internal/cli/check.go
@@ -2,7 +2,6 @@ package cli
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"io"
 	"os"
@@ -12,115 +11,6 @@ import (
 	"github.com/fatih/color"
 	"github.com/mattn/go-colorable"
 )
-
-func (a *commandApp) prepareInvocation() error {
-	i, o := &a.inv, &a.opts
-	jsonRequested := i.JSON
-	i.JSON = a.jsonOutput()
-	if !i.Legacy {
-		switch i.Operation {
-		case operationCompletion:
-			i.Shell = i.Check.Paths[0]
-		case operationRules:
-			if len(i.Check.Paths) > 0 {
-				i.Rule = i.Check.Paths[0]
-			}
-		}
-	}
-	if i.Operation != "check" && i.Render.OutputFile != "" {
-		return commandUsageError{errors.New("--output-file is only supported for checks")}
-	}
-	if i.Operation == "version" || i.Operation == "completion" {
-		return nil
-	}
-	if i.Check.Config.Disabled && i.Check.Config.Path != "" {
-		return commandUsageError{errors.New("--config cannot be combined with --no-config")}
-	}
-	if i.Operation != "check" && !i.Legacy {
-		return nil
-	}
-	explicitOutput := a.set["output-format"] || a.set["output"] || a.set["o"]
-	if i.JSON && explicitOutput && o.output != "json" && o.output != "jsonl" {
-		return commandUsageError{errors.New("--json cannot be combined with a different --output-format mode")}
-	}
-	if jsonRequested && explicitOutput && o.output == "jsonl" {
-		return commandUsageError{errors.New("--json cannot be combined with a different --output-format mode")}
-	}
-	if i.JSON && !explicitOutput {
-		o.output = "json"
-	}
-	if i.Render.Template != "" && i.Render.TemplateFile != "" {
-		return commandUsageError{errors.New("--template cannot be combined with --template-file")}
-	}
-	if (i.Render.Template != "" || i.Render.TemplateFile != "") && (i.JSON || explicitOutput) {
-		return commandUsageError{errors.New("--template/--format cannot be combined with --json or --output-format")}
-	}
-	switch o.output {
-	case "text", "oneline", "json", "jsonl", "sarif", "github":
-		if explicitOutput || i.JSON {
-			i.Render.Oneline = false
-			i.Render.Format = actionlint.OutputFormat(o.output)
-		}
-	default:
-		return commandUsageError{fmt.Errorf("invalid output mode %q: choose text, oneline, json, jsonl, sarif or github", o.output)}
-	}
-	if a.set["log-level"] {
-		switch o.logLevel {
-		case "none":
-			i.Check.Verbose, i.Check.Debug = false, false
-		case "info":
-			i.Check.Verbose, i.Check.Debug = true, false
-		case "debug":
-			i.Check.Verbose, i.Check.Debug = false, true
-		default:
-			return commandUsageError{fmt.Errorf("invalid log level %q: choose none, info or debug", o.logLevel)}
-		}
-	}
-	if o.color {
-		i.Render.Color = actionlint.ColorOptionKindAlways
-	}
-	if !i.Legacy && a.set["modern-color"] {
-		switch o.colorMode {
-		case "auto":
-			i.Render.Color = actionlint.ColorOptionKindAuto
-		case "always":
-			i.Render.Color = actionlint.ColorOptionKindAlways
-		case "never":
-			i.Render.Color = actionlint.ColorOptionKindNever
-		default:
-			return commandUsageError{fmt.Errorf("invalid color mode %q: choose auto, always or never", o.colorMode)}
-		}
-	}
-	if o.noColor {
-		i.Render.Color = actionlint.ColorOptionKindNever
-	}
-	return nil
-}
-
-func executeInvocation(ctx context.Context, streams Command, inv invocation) (int, error) {
-	switch inv.Operation {
-	case operationVersion:
-		return 0, writeVersion(streams.Stdout, inv.JSON, inv.Legacy)
-	case operationRules:
-		return 0, writeRules(streams.Stdout, inv.Rule, inv.JSON)
-	case operationDoctor:
-		return 0, writeDoctor(streams.Stdout, inv.Check, inv.JSON)
-	case operationConfigPath, operationConfigShow, operationConfigValidate:
-		return 0, runConfigCommand(streams.Stdout, inv)
-	case operationConfigInit:
-		if err := ctx.Err(); err != nil {
-			return 0, err
-		}
-		return 0, initCommandConfig(streams, inv)
-	case operationCheck:
-		if err := ctx.Err(); err != nil {
-			return 0, err
-		}
-		return executeCheck(ctx, streams, inv)
-	default:
-		return 0, commandUsageError{fmt.Errorf("unknown operation %q", inv.Operation)}
-	}
-}
 
 func executeCheck(ctx context.Context, streams Command, inv invocation) (status int, err error) {
 	c, r := inv.Check, inv.Render
@@ -255,20 +145,15 @@ func executeCheck(ctx context.Context, streams Command, inv invocation) (status 
 	return status, err
 }
 
-type commandResultWriter struct {
-	io.Writer
-	err error
-}
-
-func (w *commandResultWriter) Write(p []byte) (int, error) {
-	n, err := w.Writer.Write(p)
-	if err == nil && n < len(p) {
-		err = io.ErrShortWrite
+func analyzeCommand(app *actionlint.AnalysisSession, stdin io.Reader, paths []string, normalizeStdin bool) (*actionlint.AnalysisResult, error) {
+	switch {
+	case len(paths) == 0:
+		return app.Repository("")
+	case len(paths) == 1 && paths[0] == "-":
+		return app.ReadStdin(stdin, normalizeStdin)
+	default:
+		return app.Files(paths, nil)
 	}
-	if w.err == nil {
-		w.err = err
-	}
-	return n, err
 }
 
 func sameCommandFile(a, b string) bool {
@@ -278,20 +163,6 @@ func sameCommandFile(a, b string) bool {
 	aInfo, aErr := os.Stat(a)
 	bInfo, bErr := os.Stat(b)
 	return aErr == nil && bErr == nil && os.SameFile(aInfo, bInfo)
-}
-
-func resolveTemplate(r renderOptions) (renderOptions, error) {
-	if r.TemplateFile != "" {
-		data, err := os.ReadFile(r.TemplateFile)
-		if err != nil {
-			return r, fmt.Errorf("could not read template file %q: %w", r.TemplateFile, err)
-		}
-		r.Template = string(data)
-		if r.Template == "" {
-			return r, fmt.Errorf("template file %q is empty", r.TemplateFile)
-		}
-	}
-	return r, nil
 }
 
 func absPath(path string) string {
